+++ src/components/Questionnaire/Description.jsx
@@ -1,0 +1,15 @@
+import "../..//styles/questionnaire/Description.css";
+
+const Description = () => {
+    return (
+    <div className="DescriptionMent">
+      <h1 className="Title">상대방에 대해 알려주세요</h1>
+      <p className="Description">
+        선택하신 정보로 선물을 추천해드립니다.<br />
+        다른 목적으로 사용되지 않으며 제 3자에게 제공되지 않습니다.
+      </p>
+      </div>
+      );
+      };
+
+      export default Description
