+++ src/components/Questionnaire/SelectionButton.jsx
@@ -1,0 +1,40 @@
+import MaleIcon from "../icons/male_icon.svg";
+import FemaleIcon from "../icons/female_icon.svg";
+import { useUser } from "../contexts/UserContext";
+import "../../styles/questionnaire/SelectionButton.css";
+
+const SelectionButton = () => {
+    const { userData, setUserData } = useUser(); 
+
+    return (
+    <div>
+      <div className="SelectionGroup">
+        <div className="ButtonGroup">
+          <button className={`SelectionButton gender ${userData.gender === "남자" ? "selected" : ""}`}
+                  onClick={() => setUserData(prev => ({ ...prev, gender: "남자" }))}>
+                     <img src={MaleIcon} alt="남자 아이콘" width={92} height={92} />
+                     남자
+                     </button>
+          <button className={`SelectionButton gender ${userData.gender === "여자" ? "selected" : ""}`}
+                  onClick={() => setUserData(prev => ({ ...prev, gender: "여자" }))}>
+                    <img src={FemaleIcon} alt="여자 아이콘" width={92} height={92} />
+                    여자
+                    </button>
+        </div>
+      </div>
+
+      <div className="SelectionGroup">
+        <div className="ButtonGroup">
+          {["부모님", "애인", "친구"].map(item => (
+            <button key={item} className={`SelectionButton relationship ${userData.relationship === item ? "selected" : ""}`}
+                    onClick={() => setUserData(prev => ({ ...prev, relationship: item }))}>
+              {item}
+            </button>
+          ))}
+        </div>
+      </div>
+      </div>
+    );
+};
+
+export default SelectionButton;
