+++ src/components/common/Footer.css
@@ -1,0 +1,7 @@
+.Footer {
+    width: 360px;
+    height: 35px;
+    background-color: #333;
+    position: absolute;
+    bottom: 0;
+  }
