+++ src/components/common/Footer.jsx
@@ -1,0 +1,9 @@
+import "../common/Footer.css";
+
+const Footer = () => {
+return(
+    <div className="Footer"></div>
+);
+};
+
+export default Footer;
