+++ src/styles/GiftThemeSelection/GiftThemeSelection.css
@@ -2,14 +2,14 @@
   width: 360px;
   height: 740px; /* 강제 높이 지정 */
   min-height: 740px; /* 최소 높이 설정 */
+  font-family: "Pretendard";
   display: flex;
-  flex-direction: column;
-  justify-content: space-between; /* 하단 여백 확보 */
+  align-content: center;
+  flex-direction: row;
+  justify-content: center;
   align-items: center;
-  text-align: center;
-  padding: 20px;
+  flex-wrap: wrap;
   background-color: #fff;
-  position: relative;
 }
