+++ src/styles/onboarding/SlideImage.css
@@ -1,0 +1,17 @@
+/* Swiper 슬라이드 스타일 */
+.swiper-slide {
+    width: 360px !important;
+    max-width: 360px !important;
+    height: 451px !important;
+    background-color: #D9D9D9 !important;
+    display: flex !important;
+    align-items: center !important;
+    justify-content: center !important;
+    margin-top: 95px !important;
+  }
+  
+  /* Swiper 래퍼 */
+  .swiper-wrapper {
+    width: 100% !important;
+    max-width: 360px !important;
+  }
