+++ src/styles/onboarding/StartButton.css
@@ -1,0 +1,7 @@
+/* 버튼 컨테이너 */
+.StartButtonContainer {
+    margin-top: 100px; /* 버튼 위쪽 여백 줄임 */
+    margin-bottom: 0px; /* 버튼 아래 여백도 조정 */
+    display: flex;
+    justify-content: center;
+  }
