+++ src/styles/questionnaire/NextButton.css
@@ -1,0 +1,5 @@
+    .NextButtonContainer {
+        margin-bottom: 25px;
+        display: flex;
+        justify-content: center;
+      }
